Added PID for TP-Link TL-WN821N (usb)
For some reason, until kernel 4.14 this driver was working properly
but since last update to 4.15 the TL-WN821N was not being recognized.
Checking with lshw displayed that Realtek device was UNCLAIMED.
To solve this problem, I added PID 2357 VID 0107 to usb_intf.c list
and it start to work again.